+++ resources/[dg]/dg-gangs/typescript/src/server/controllers/events.ts
@@ -2,33 +2,6 @@ import { Chat, Core, RPC, Util } from '@dgx/server';
 import gangManager from 'classes/gangmanager';
 import { dispatchCurrentGangToClient } from 'helpers';
 
-Chat.registerCommand(
-  'createGang',
-  'Create a new gang',
-  [
-    {
-      name: 'name',
-      description: 'Name of gang',
-    },
-    {
-      name: 'label',
-      description: 'label of gang',
-    },
-    {
-      name: 'cid',
-      description: 'CitizenID of owner',
-    },
-  ],
-  'developer',
-  (src, _, params) => {
-    const cid = Number(params[2]);
-    if (isNaN(cid)) {
-      throw new Error('CitizenId should be a valid integer');
-    }
-    gangManager.createGang(params[0], params[1], cid);
-  }
-);
-
 RPC.register('gangs:server:getData', async (plyId: number): Promise<Gangs.Data | null> => {
   const cid = Util.getCID(plyId);
   const gang = gangManager.getPlayerGang(cid);