+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optionele locatie knop toegevoegd aan phone mails
 - Nieuw guard spawning systeem toegevoegd om makkelijker en beter enemy NPCs te spawnen
 - Nieuwe criminele drugs activiteit toegevoegd
+- Commands toegevoegd om gangs te creeren/verwijderen
 
 ### Changed
 - Override CID entry toegevoegd om kofi vehicles te kunnen assignen aan offline vehicles